requirements.txt
Django==1.8.4
djangorestframework==3.2.4
Markdown==2.6.2
wheel==0.24.0
Pillow==3.0.0